requirements.txt
# FastAPI and web dependencies
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
starlette>=0.27.0
httpx>=0.25.0
python-multipart>=0.0.6
requests>=2.31.0

# LangChain and AI dependencies
langchain>=0.0.267
langchain-community>=0.0.10
langchain-google-genai>=0.0.3
langchain-groq>=0.0.1
faiss-cpu>=1.7.4
pypdf>=3.15.0

# PDF and file handling
reportlab>=3.6.0
python-magic>=0.4.24

# Database and storage
sqlalchemy>=2.0.0
redis>=4.5.1
alembic>=1.11.1

# Security and authentication
passlib>=1.7.4
python-jose>=3.3.0
bcrypt>=4.0.1

# Monitoring and observability
prometheus-client>=0.16.0
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
opentelemetry-exporter-otlp>=1.18.0
opentelemetry-instrumentation-fastapi>=0.39b0
gunicorn>=21.2.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
coverage>=7.2.7